requirements.txt
Django==4.2.16
djangorestframework==3.16.1
psycopg2-binary==2.9.9
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
python-decouple==3.8
numpy==2.1.2
scipy==1.14.1
weasyprint==62.3
Pillow==10.4.0
gunicorn==23.0.0